(SERVER-2265) Refactor: move file system utils
