+++ lib/puppetserver/ca/action/import.rb
@@ -1,5 +1,5 @@
 require 'optparse'
-require 'puppetserver/utils/file_utilities'
+require 'puppetserver/ca/utils/file_system'
 require 'puppetserver/ca/x509_loader'
 require 'puppetserver/ca/puppet_config'
 require 'puppetserver/ca/utils'
