+++ lib/puppetserver/ca/utils/file_system.rb
